🐛Bug fix
When the UI for the search results was modified, the method for displaying the title was changed from using an h1 tag to using a div tag. However, because the query selector in the JavaScript code for changing the title was not updated, there was a bug that prevented the title from being changed.